docker/rng_client/requirements.txt
paho-mqtt
dnspython3